pyproject.toml
[project]
name = "font-fix"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]
